Don’t break idempotency if vid is declared as string
In modules_utils/netbox_ipam.py _build_diff():
When an already present vlan with the same data is pushed, there’s a
diff between vid: before is a int and after is a string.
Changed = True if diff is instanced so every values in before
and after are converted to string in order to compare them. If they are equals,
diff = None.