Add a 3rd test environment - latest netbox with latest ansible (#144)
The previous test environments:
* Previous netbox, latest ansible
* Latest netbox, dev ansible

This left the most-used combination of latest netbox with latest ansible untested.

New test environments:
* Previous netbox, latest ansible
* Latest netbox, latest ansible
* Dev netbox, dev ansible

This way if the dev ansible is broken for any reason we're still getting a test against the latest netbox. Also introduces testing against the dev version of netbox so we can catch any potential issues there.