Enable faster lookups on Netbox version < 3.0.6 (#645)
* Enable faster lookups on Netbox version < 3.0.6

This creates a new code path where we attempt to lookup the interfaces,
and then trace them back via the cable attribute to find the exact
match. This is to help performance on older versions of NetBox where the
GET api does not take termination_*_id parameters to provide an exact
match.

The old behavior of looping through ALL the cables is preserved in the
case where an exact match is not found, however it probably should be
removed.