Fixes #943: Partial update of custom fields is not idempotent
It is important that the object's custom fields dictionary contains only those keys that were given by the user. Otherwise the following dictionary comparison does not work and leads to the result, that there is a difference in state even when the entries already have the desired state. The issue is caused, because the comparison takes also those entries into account that were not given by the user.

Signed-off-by: Christoph Fiehe <c.fiehe@eurodata.de>