Handle netbox api cache more gracefully
There are instances where we fail to read the cached
netbox_api_dump.json file, and the code was setting openapi to an empty
dictionary, and netbox_api_version to 0.

This is not correct, because we can still fetch the netbox_api_version
by querying the live server. In addition, in cases where the cached API was
failing to read, it would end up triggering the if/else statement to try
and access the older API doc endpoint, even if the server was not <
3.5.0

Overall the code was very convoluted and could break if the cached file
was not readable or writable. This patch cleans the logic up and makes
it more clear what is going on.