feat: Remove netbox_tunnel.yml from integration tests
This commit removes the `netbox_tunnel.yml` file from the integration tests for targets v3.5 and v3.6. The file was previously included using the `ansible.builtin.include_tasks` module, but it is no longer needed. This change ensures that the `netbox_tunnel` tests are no longer executed as part of the integration test suite.