Allow dict parameters for primary_mac_address
Coercing primary_mac_address to str prevents using dict constructs to
point to a specific object in netbox, instead always searching for an
entry with the plaintext mac.
This key for the object is not guaranteed to be unique, resulting in
errors when multiple objects are returned.

Allowing raw parameters permits the following construct to lookup mac
addresses in a different way to get an unique object, for example by id:
```yaml
netbox_device_interface:
  primary_mac_address:
    id: "{{ mac_address_object_id }}"
```

Fixes: #1413