ci: Fix release workflow running too often
on.create triggers on any tag *or branch* creation, and doesn't have a
filter `on.push.tags`.
(https://docs.github.com/en/actions/reference/workflows-and-actions/events-that-trigger-workflows#create)

The correct way to only run a job when a tag matching a pattern is
pushed is via `on.push.tags`.
(https://docs.github.com/en/actions/reference/workflows-and-actions/events-that-trigger-workflows#running-your-workflow-only-when-a-push-of-specific-tags-occurs)